Adding autocomplete to post, community, message, and site forms. Fixes #453
